pyproject.toml
[project]
name = "discord-tldr"
version = "0.1.0"
description = "A Discord bot that generates TLDR summaries of channel messages using OpenAI"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "discord.py>=2.3.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
]


